app logo and icon

+++ lib/widgets/3d_model_viewer.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:model_viewer_plus/model_viewer_plus.dart';
 import 'package:studio_yaiverse_mobile/services/api_service.dart';
+import 'package:studio_yaiverse_mobile/views/home_view.dart';
 
 class ThreeDModelViewer extends StatelessWidget {
   final String name, username;
@@ -48,12 +49,48 @@ class ThreeDModelViewer extends StatelessWidget {
                 right: 80,
                 child: GestureDetector(
                   onTap: () {
-                    ApiService.delete(username, name);
+                    showDialog<String>(
+                        context: context,
+                        builder: (BuildContext context) => AlertDialog(
+                                content: const Text('모델을 삭제할까요?'),
+                                actions: <Widget>[
+                                  TextButton(
+                                    onPressed: () => {
+                                      ApiService.delete(username, name),
+                                      Navigator.pushAndRemoveUntil(
+                                          context,
+                                          MaterialPageRoute(
+                                            builder: (context) =>
+                                                const StudioYaiverseHome(
+                                                    title: "Studio YAIverse"),
+                                          ),
+                                          (route) => false)
+                                    },
+                                    child: const Text(
+                                      '확인',
+                                      style:
+                                          TextStyle(color: Color(0xffBB2649)),
+                                    ),
+                                  ),
+                                  TextButton(
+                                    onPressed: () =>
+                                        Navigator.of(context).pop(),
+                                    child: const Text(
+                                      '취소',
+                                      style: TextStyle(color: Colors.grey),
+                                    ),
+                                  )
+                                ]));
                   },
-                  child: ClipRect(
-                    child: Container(
-                      decoration: const BoxDecoration(color: Colors.white),
-                      child: const Text("Delete"),
+                  child: Container(
+                    margin: const EdgeInsets.all(20),
+                    padding: const EdgeInsets.all(10),
+                    decoration: BoxDecoration(
+                        borderRadius: BorderRadius.circular(100),
+                        border: Border.all(width: 2, color: Colors.white)),
+                    child: const Text(
+                      "Delete",
+                      style: TextStyle(color: Colors.white),
                     ),
                   ),
                 )),
